Pipe from player to main now from player to output.
Many, many, changes.  Lots of restructuring of
signal, threads, pipes.  Big change.  It was
problematic for player to send signal to main
then for main to echo the signal to output.
It is better for there to be a pipe directly from
player to output instead of a pipe from player
to main.